agent/requirements.txt
livekit>=0.17.0
livekit-agents>=0.9.0
livekit-plugins-openai>=0.7.0
livekit-plugins-deepgram>=0.7.0
livekit-plugins-cartesia>=0.3.0
livekit-plugins-silero>=0.7.0
openai>=1.54.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
